+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Cerulean City now uses its latest visual layout.
 - Pallet Town and Cerulean Gym now use their latest visual layouts.
 - Route 24 now uses its latest expanded visual layout.
+- Route 25 now uses its complete visual layout.
 
 **Fixed**